Issue-3 || Removing the performance metrics from the api model and moving it to radar model as planned and renamed some files as well to avoid confusion
